Fix lint error with strtol in test
Review URL: http://codereview.chromium.org/549111

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@3666 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/test-debug.cc
@@ -3881,11 +3881,11 @@ int GetTotalFramesInt(char *message) {
   }
   pos += strlen(prefix);
   char* pos_end = pos;
-  long res = strtol(pos, &pos_end, 10);
+  int res = static_cast<int>(strtol(pos, &pos_end, 10));
   if (pos_end == pos) {
     return -1;
   }
-  return static_cast<int>(res);
+  return res;
 }
